requirements.txt
PyQt5==5.15.10
av==11.0.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
